Split native Zenmap result action helpers
